+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Add support for multiple workers
+
 ### 0.2.1 - 2019-11-07
 
  - Fix requirements